Added Backup and Restore functions
The backup and restore functions are handled by the HelperFunctions class.

Backup saves the list of items and any dayupdates it finds in the working directory into a .stockout file, which is just a JSON file with 2 serialised lists seperated by the Registered Trademark character.

Restore parses a selected backup file and returns a list of Items and GridEntries (read from all the dayupdate files).

Additionally, the SaveRestoredDayUpdates method goes through the list of GridEntries then overwrites any Existing JSON files for those days